fix examples/sample.exs
The issue was, that the examples used a binary for file name in
the usage of hdr_histogram:log/3.

Since the usage of this library is not elixir specific, I took the
liberty of removing the example from the README.

I also applied common elixir code styles to the best of my knowledge.