Support Mac M1 / arm64 architecture
This commit updates the Makefile for the C source code to check
the architecture via `uname -p` before initializing the compiler
flags. Mac M1s will report `arm` architecture, and in that case
we change the cc flags from `x86_64` to `arm64`. Prior to this
patch, trying to use this library on an m1 causes an error.
With the patch, it runs properly.